chore: move more types to react package
